@@ -1,31 +1,29 @@
-﻿using MusicStoreUI.Services;
-using MusicStoreUI.Models;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Steeltoe.Discovery.Client;
-
-#if USE_REDIS_CACHE
-using Microsoft.AspNetCore.DataProtection;
-using Steeltoe.CloudFoundry.Connector.Redis;
-using Steeltoe.Security.DataProtection;
-#endif
-
-using Steeltoe.CloudFoundry.Connector.MySql.EFCore;
+using Microsoft.Extensions.Hosting;
+using MusicStoreUI.Models;
+using MusicStoreUI.Services;
 using Steeltoe.CircuitBreaker.Hystrix;
-using Command = MusicStoreUI.Services.HystrixCommands;
-using Steeltoe.CloudFoundry.Connector;
+using Steeltoe.Common.Http.Discovery;
+using Steeltoe.Connector.MySql.EFCore;
+using Steeltoe.Connector.Redis;
+using Steeltoe.Discovery.Client;
 using Steeltoe.Management.Endpoint.Health;
 using Steeltoe.Management.Endpoint.Hypermedia;
 using Steeltoe.Management.Endpoint.Info;
 using Steeltoe.Management.Endpoint.Loggers;
 using Steeltoe.Management.Endpoint.Mappings;
+using Steeltoe.Management.Endpoint.SpringBootAdminClient;
 using Steeltoe.Management.Endpoint.Trace;
-using Microsoft.EntityFrameworkCore;
+using Steeltoe.Security.DataProtection;
 using System;
+using Command = MusicStoreUI.Services.HystrixCommands;
 
 namespace MusicStoreUI
 {
@@ -41,20 +39,15 @@ namespace MusicStoreUI
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            var connMgr = new ConnectionStringManager(Configuration);
-            
+
             // Add framework services.
-#if USE_REDIS_CACHE
             services.AddRedisConnectionMultiplexer(Configuration);
             services.AddDataProtection()
                 .PersistKeysToRedis()
                 .SetApplicationName("MusicStoreUI");
             services.AddDistributedRedisCache(Configuration);
-#else
-            services.AddDistributedMemoryCache();
-#endif
 
-            // Add managment endpoint services
+            // Add management endpoint services
             services.AddHypermediaActuator(Configuration);
             services.AddInfoActuator(Configuration);
             services.AddHealthActuator(Configuration);
@@ -65,34 +58,22 @@ namespace MusicStoreUI
             services.Configure<AppSettings>(Configuration.GetSection("AppSettings"));
 
             services.AddDbContext<AccountsContext>(options => options.UseMySql(Configuration));
-            services.ConfigureApplicationCookie(options => options.AccessDeniedPath = "/Home/AccessDenied");
             services.AddIdentity<ApplicationUser, ApplicationRole>()
                     .AddEntityFrameworkStores<AccountsContext>()
                     .AddDefaultTokenProviders();
-            services.ConfigureApplicationCookie(options => options.LoginPath = "/Account/LogIn");
 
-            services.AddDiscoveryClient(Configuration);
-
-            services.AddSingleton<IMusicStore, MusicStoreService>();
-            services.AddSingleton<IShoppingCart, ShoppingCartService>();
-            services.AddSingleton<IOrderProcessing, OrderProcessingService>();
-
-            services.AddHystrixCommand<Command.GetTopAlbums>("MusicStore", Configuration);
-            services.AddHystrixCommand<Command.GetGenres>("MusicStore", Configuration);
-            services.AddHystrixCommand<Command.GetGenre>("MusicStore", Configuration);
-            services.AddHystrixCommand<Command.GetAlbum>("MusicStore", Configuration);
-
-            services.AddMvc();
-
-            // Add memory cache services
-            services.AddMemoryCache();
-
-            // Add session related services.
-
-            // Use call below if you want sticky Sessions on Cloud Foundry
-            // services.AddSession((options) => options.CookieName = "JSESSIONID");
-
-            services.AddSession();
+            services.ConfigureApplicationCookie(options =>
+            {
+                options.AccessDeniedPath = "/Home/AccessDenied";
+                options.LoginPath = "/Account/LogIn";
+                options.SlidingExpiration = true;
+            });
+            services.AddSession(options =>
+            {
+                options.IdleTimeout = TimeSpan.FromMinutes(20);
+                options.Cookie.HttpOnly = false;
+                options.Cookie.IsEssential = true;
+            });
 
             // Configure Auth
             services.AddAuthorization(options =>
@@ -106,12 +87,25 @@ namespace MusicStoreUI
                     });
             });
 
+            services.AddDiscoveryClient(Configuration);
+
+            services.AddHttpClient<IMusicStore, MusicStoreService>().AddRoundRobinLoadBalancer();
+            services.AddHttpClient<IShoppingCart, ShoppingCartService>().AddServiceDiscovery();
+            services.AddHttpClient<IOrderProcessing, OrderProcessingService>().AddServiceDiscovery();
+
+            services.AddHystrixCommand<Command.GetTopAlbums>("MusicStore", Configuration);
+            services.AddHystrixCommand<Command.GetGenres>("MusicStore", Configuration);
+            services.AddHystrixCommand<Command.GetGenre>("MusicStore", Configuration);
+            services.AddHystrixCommand<Command.GetAlbum>("MusicStore", Configuration);
+
+            services.AddControllersWithViews();
+
             // Add Hystrix metrics stream to enable monitoring
             services.AddHystrixMetricsStream(Configuration);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             // Add Hystrix Metrics context to pipeline
             app.UseHystrixRequestContext();
@@ -125,7 +119,7 @@ namespace MusicStoreUI
             app.UseMappingsActuator();
 
             app.UseSession();
-
+            
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -136,22 +130,25 @@ namespace MusicStoreUI
             }
 
             app.UseStaticFiles();
+            app.UseRouting();
 
             // Add cookie-based authentication to the request pipeline
             app.UseAuthentication();
+            app.UseAuthorization();
 
-            app.UseMvc(routes =>
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "areaRoute",
-                    template: "{area:exists}/{controller}/{action}",
+                    pattern: "{area:exists}/{controller}/{action}",
                     defaults: new { action = "Index" });
 
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
 
+            app.RegisterSpringBootAdmin(Configuration);
             app.UseDiscoveryClient();
 
             // Startup Hystrix metrics stream
