@@ -1,6 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using Steeltoe.Common.Discovery;
 using System;
 using System.IO;
 using System.Net;
@@ -12,93 +11,73 @@ namespace MusicStoreUI.Services
 {
     public abstract class BaseDiscoveryService
     {
-        protected DiscoveryHttpClientHandler _handler;
+        protected HttpClient _client;
         protected ILogger _logger;
 
-        public BaseDiscoveryService(IDiscoveryClient client, ILogger logger)
+        public BaseDiscoveryService(HttpClient client, ILogger logger)
         {
-            _handler = new DiscoveryHttpClientHandler(client, logger);
+            _client = client;
             _logger = logger;
         }
-        public virtual HttpClient GetClient()
-        {
-            var client = new HttpClient(_handler, false);
-            return client;
-        }
+
         public virtual async Task<bool> Invoke(HttpRequestMessage request)
         {
-            var client = GetClient();
             try
             {
-                using (HttpResponseMessage response = await client.SendAsync(request))
-                {
-                    var stream = await response.Content.ReadAsStreamAsync();
-                    return response.StatusCode == HttpStatusCode.OK;
-                }
+                using var response = await _client.SendAsync(request);
+                var stream = await response.Content.ReadAsStreamAsync();
+                return response.StatusCode == HttpStatusCode.OK;
             }
             catch (Exception e)
             {
                 _logger?.LogError("Invoke exception: {0}", e);
                 throw;
             }
-
-            //return false;
         }
         public virtual async Task<bool> Invoke(HttpRequestMessage request, object content)
         {
-            var client = GetClient();
             try
             {
                 request.Content = Serialize(content);
-                using (HttpResponseMessage response = await client.SendAsync(request))
-                {
-                    var stream = await response.Content.ReadAsStreamAsync();
-                    return response.StatusCode == HttpStatusCode.OK;
-                }
+                using var response = await _client.SendAsync(request);
+                var stream = await response.Content.ReadAsStreamAsync();
+                return response.StatusCode == HttpStatusCode.OK;
             }
             catch (Exception e)
             {
                 _logger?.LogError("Invoke exception: {0}", e);
                 throw;
             }
-
-            //return false;
         }
 
         public virtual async Task<T> Invoke<T>(HttpRequestMessage request, object content = null)
         {
-            var client = GetClient();
             try
             {
                 if (content != null)
                 {
                     request.Content = Serialize(content);
                 }
-                using (HttpResponseMessage response = await client.SendAsync(request))
-                {
-                    var stream = await response.Content.ReadAsStreamAsync();
-                    return Deserialize<T>(stream);
-                }
+                using var response = await _client.SendAsync(request);
+                var stream = await response.Content.ReadAsStreamAsync();
+                return Deserialize<T>(stream);
             }
             catch (Exception e)
             {
                 _logger?.LogError("Invoke exception: {0}", e);
                 throw;
             }
-
-            //return default(T);
         }
 
         public virtual T Deserialize<T>(Stream stream)
         {
             try
             {
-                using (JsonReader reader = new JsonTextReader(new StreamReader(stream)))
-                {
-                    JsonSerializer serializer = new JsonSerializer();
-                    return (T)serializer.Deserialize(reader, typeof(T));
-                }
-            } catch(Exception e)
+                using JsonReader reader = new JsonTextReader(new StreamReader(stream));
+                var serializer = new JsonSerializer();
+                return (T)serializer.Deserialize(reader, typeof(T));
+            }
+            catch(Exception e)
             {
                 _logger?.LogError("Deserialize exception: {0}", e);
                 throw;
@@ -108,7 +87,7 @@ namespace MusicStoreUI.Services
 
         public virtual HttpContent Serialize(object toSerialize)
         {
-            string json = JsonConvert.SerializeObject(toSerialize);
+            var json = JsonConvert.SerializeObject(toSerialize);
             return new StringContent(json, Encoding.UTF8, "application/json");
         }
     }
