@@ -1,10 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿using Microsoft.Extensions.Logging;
 using MusicStoreUI.Models;
-using Microsoft.Extensions.Logging;
+using System;
+using System.Collections.Generic;
 using System.Net.Http;
-using Steeltoe.Common.Discovery;
+using System.Threading.Tasks;
 
 namespace MusicStoreUI.Services
 {
@@ -12,10 +11,12 @@ namespace MusicStoreUI.Services
     {
         private const string SHOPPINGCART_URL = "http://shoppingcart/api/ShoppingCart/{cartId}";
         private const string SHOPPINGCART_ITEM_URL ="http://shoppingcart/api/ShoppingCart/{cartId}/Item/{itemId}";
+        private new ILogger _logger;
 
-        public ShoppingCartService(IDiscoveryClient client, ILoggerFactory logFactory) :
+        public ShoppingCartService(HttpClient client, ILoggerFactory logFactory) :
             base(client, logFactory.CreateLogger<ShoppingCartService>())
         {
+            _logger = logFactory.CreateLogger<ShoppingCartService>();
         }
 
         public async Task<bool> EmptyCartAsync(string cartId)
@@ -32,7 +33,15 @@ namespace MusicStoreUI.Services
             var cartUrl = SHOPPINGCART_URL.Replace("{cartId}", cartId);
 
             var request = new HttpRequestMessage(HttpMethod.Get, cartUrl);
-            var cartResult = await Invoke<List<CartItemJson>>(request);
+            var cartResult = new List<CartItemJson>();
+            try
+            {
+                cartResult = await Invoke<List<CartItemJson>>(request);
+            }
+            catch (Exception e)
+            {
+                _logger.LogCritical(e, "Failed to get cart!");
+            }
 
             var result = CartItem.From(cartResult);
             return result;
@@ -61,8 +70,16 @@ namespace MusicStoreUI.Services
             var cartUrl = SHOPPINGCART_URL.Replace("{cartId}", cartId);
 
             var request = new HttpRequestMessage(HttpMethod.Put, cartUrl);
-            var result = await Invoke(request);
-            return result;
+            try
+            {
+                var result = await Invoke(request);
+                return result;
+            }
+            catch (Exception e)
+            {
+                _logger.LogCritical(e, "Failed to create cart!");
+                return false;
+            }
         }
     }
 }
