@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using ShoppingCartService.Models;
-
-using Microsoft.EntityFrameworkCore;
-
-// For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace ShoppingCartService.Controllers
 {
@@ -25,16 +21,14 @@ namespace ShoppingCartService.Controllers
 
         public ShoppingCartContext DbContext { get; }
 
-        //
-        // GET: api/ShoppingCart/cartId
-        [HttpGet("{cartId}")]
-        public async Task<IActionResult> GetCartItems(string cartId)
+        // GET: api/ShoppingCart/id
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetCartItems(string id)
         {
-    
+            _logger?.LogTrace("Get cart {cartId}", id);
             var cart = await DbContext.Carts
-                .Where(c => c.CartId == cartId)
                 .Include(g => g.CartItems)
-                .FirstOrDefaultAsync();
+                .FirstOrDefaultAsync(c => c.CartId == id);
 
             if (cart == null)
             {
@@ -45,13 +39,12 @@ namespace ShoppingCartService.Controllers
             return new ObjectResult(result);
         }
 
-        // PUT: api/ShoppingCart/cartid
-        [HttpPut("{cartId}")]
-        public async Task<IActionResult> CreateCart(string cartId)
+        // PUT: api/ShoppingCart/id
+        [HttpPut("{id}")]
+        public async Task<IActionResult> CreateCart(string id)
         {
-            var cart = await DbContext.Carts
-                        .Where(c => c.CartId == cartId)
-                        .FirstOrDefaultAsync();
+            _logger?.LogTrace("Create cart {cartId}", id);
+            var cart = await DbContext.Carts.FirstOrDefaultAsync(c => c.CartId == id);
 
             if (cart != null)
             {
@@ -59,22 +52,21 @@ namespace ShoppingCartService.Controllers
             }
             cart = new ShoppingCart()
             {
-                CartId = cartId
-
+                CartId = id
             };
             DbContext.Carts.Add(cart);
             await DbContext.SaveChangesAsync();
             return Ok();
         }
 
-        // DELETE: api/ShoppingCart/cartid
-        [HttpDelete("{cartId}")]
-        public async Task<IActionResult> DeleteCart(string cartId)
+        // DELETE: api/ShoppingCart/id
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteCart(string id)
         {
+            _logger?.LogTrace("Delete cart {cartId}", id);
             var cart = await DbContext.Carts
                         .Include(c => c.CartItems)
-                        .Where(c => c.CartId == cartId)
-                        .FirstOrDefaultAsync();
+                        .FirstOrDefaultAsync(c => c.CartId == id);
 
             if (cart == null)
             {
@@ -87,24 +79,22 @@ namespace ShoppingCartService.Controllers
             return Ok();
         }
 
-        //
-        // PUT: api/ShoppingCart/cartid/Item/itemId
-        [HttpPut("{cartId}/Item/{itemId}")]
-        public async Task<IActionResult> AddCartItem(string cartId, int itemId)
+        // PUT: api/ShoppingCart/id/Item/itemId
+        [HttpPut("{id}/Item/{itemId}")]
+        public async Task<IActionResult> AddCartItem(string id, int itemId)
         {
+            _logger?.LogTrace("Add Item {itemId} to cart {id}", itemId, id);
             var cart = await DbContext.Carts
-                       .Where(c => c.CartId == cartId)
                        .Include(g => g.CartItems)
-                       .FirstOrDefaultAsync();
+                       .FirstOrDefaultAsync(c => c.CartId == id);
 
             if (cart == null)
             {
+                _logger?.LogCritical("Cart not found!");
                 return NotFound();
             }
 
-            var cartItem = cart.CartItems
-                .Where(item => item.ItemKey == itemId)
-                .SingleOrDefault();
+            var cartItem = cart.CartItems.SingleOrDefault(item => item.ItemKey == itemId);
 
             if (cartItem == null)
             {
@@ -112,13 +102,14 @@ namespace ShoppingCartService.Controllers
                 cartItem = new CartItem
                 {
                     ItemKey = itemId,
-                    CartId = cartId,
+                    CartId = id,
                     Count = 1,
                     DateCreated = DateTime.Now
                 };
 
-                DbContext.CartItems.Add(cartItem);
-            } else
+                cart.CartItems.Add(cartItem);
+            }
+            else
             {
                 cartItem.Count++;
             }
@@ -126,28 +117,27 @@ namespace ShoppingCartService.Controllers
             return Ok();
         }
 
-        //
         // DELETE: /api/ShoppingCart/{cartId}/Item/itemId
-        [HttpDelete("{cartId}/Item/{itemId}")]
+        [HttpDelete("{id}/Item/{itemId}")]
         //[ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteCartItem(string cartId, int itemId)
+        public async Task<IActionResult> DeleteCartItem(string id, int itemId)
         {
+            _logger?.LogTrace("Remove Item {itemId} from cart {cartId}", itemId, id);
             var cart = await DbContext.Carts
-                            .Where(c => c.CartId == cartId)
                             .Include(g => g.CartItems)
-                            .FirstOrDefaultAsync();
+                            .FirstOrDefaultAsync(c => c.CartId == id);
 
             if (cart == null)
             {
+                _logger?.LogCritical("Cart not found!");
                 return NotFound();
             }
 
-            var cartItem = cart.CartItems
-                .Where(item => item.ItemKey == itemId)
-                .SingleOrDefault();
+            var cartItem = cart.CartItems.SingleOrDefault(item => item.ItemKey == itemId);
 
             if (cartItem == null)
             {
+                _logger?.LogCritical("Cart item not found!");
                 return NotFound();
             }
 
